Segregate mainnet chainhandler tests and regtest chainhandler tests
